+++ src/filters/RedirectFilter.php
@@ -8,8 +8,6 @@
 
 namespace flipbox\ember\filters;
 
-use Craft;
-use craft\helpers\ArrayHelper;
 use flipbox\ember\helpers\ControllerHelper;
 use yii\base\Action;
 use yii\base\ActionFilter;
@@ -24,50 +22,14 @@ use yii\web\Response;
  */
 class RedirectFilter extends ActionFilter
 {
-    /**
-     * @var array this property defines the status code mapping for each action.
-     * For each action that should only support limited set of status codes
-     * you add a status code with the action id as array key and an array value of
-     * allowed status codes (e.g. 'create' => [201, 204], 'delete' => [204]).
-     * If an action is not defined the default statusCode property will be used.
-     *
-     * You can use `'*'` to stand for all actions. When an action is explicitly
-     * specified, it takes precedence over the specification given by `'*'`.
-     *
-     * For example,
-     *
-     * ```php
-     * [
-     *   'create' => [201, 204],
-     *   'update' => [200],
-     *   'delete' => [204],
-     *   '*' => [200],
-     * ]
-     * ```
-     */
-    public $actions = [];
-
-    /**
-     * Allow redirection of a null result
-     * @var bool
-     */
-    public $statusCode = null;
+    use traits\ActionTrait,
+        traits\FormatTrait;
 
     /**
      * Allow redirection of a null result
      * @var bool
      */
     public $allowNull = false;
-
-    /**
-     * @var string
-     */
-    public $format = Response::FORMAT_RAW;
-
-    /**
-     * @var array
-     */
-    public $formats = [];
 
     /**
      * @param \yii\base\Action $action
@@ -77,7 +39,7 @@ class RedirectFilter extends ActionFilter
     public function afterAction($action, $result)
     {
         if ($this->formatMatch($action->id) &&
-            $this->statusCodeMatch($action->id) &&
+            $this->actionMatch($action->id) &&
             $this->resultMatch($result)
         ) {
             return $this->redirect($action, $result);
@@ -106,65 +68,5 @@ class RedirectFilter extends ActionFilter
     protected function resultMatch($result): bool
     {
         return $result !== null || ($this->allowNull === true);
-    }
-
-    /**
-     * @param string $action
-     * @return bool
-     */
-    protected function statusCodeMatch(string $action): bool
-    {
-        if (!$statusCode = $this->findStatusCode($action)) {
-            return true;
-        }
-
-        return Craft::$app->getResponse()->getStatusCode() === $statusCode;
-    }
-
-    /**
-     * @param string $action
-     * @return string|null
-     */
-    protected function findStatusCode(string $action)
-    {
-        // Default status code
-        $statusCode = ArrayHelper::getValue($this->actions, '*', $this->actions);
-
-        // Look for definitions
-        if (isset($this->actions[$action])) {
-            $statusCode = $this->actions[$action];
-        }
-
-        return $statusCode;
-    }
-
-    /**
-     * @param string $action
-     * @return bool
-     */
-    protected function formatMatch(string $action): bool
-    {
-        if (!$format = $this->findFormat($action)) {
-            return true;
-        }
-
-        return Craft::$app->getResponse()->format === $format;
-    }
-
-    /**
-     * @param string $action
-     * @return string|null
-     */
-    protected function findFormat(string $action)
-    {
-        // Default format
-        $format = ArrayHelper::getValue($this->formats, '*', $this->format);
-
-        // Look for definitions
-        if (isset($this->formats[$action])) {
-            $format = $this->formats[$action];
-        }
-
-        return $format;
     }
 }