+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+## Unreleased
+### Added
+- ModelError filter
+- Filter traits
+
 ## 1.0.0-rc.3 - 2017-12-09
 ### Added
 - FlashMessages filter